Implement first ergonomic constructor, in SelectBitImageMode
Instantiation of the subclass in CmdBase's static method `from()` is
changed to use `Object.create()` so as to continue to use the subclass's
prototype while bypassing its constructor.

Member validation is factored out into an instance method on CmdBase so
that it can be called both in its static method `from()` and in its
constructor.