Fix ctors after retargeting from ES5 to es2022
Had to retarget to es2022 to make use of the new stage 3 decorator and
decorator metadata API.

Unfortunately, after ES5, it is no longer possible for the parent ctor
to assign to fields declared on the child class. I was using this
"feature" of ES5 to ensure validate() was called at the end of
construction, but it's not possible anymore, so I'm just having the
child ctor call validate() manually. Not ideal.