Major refactor: move parsing, validation logic into decorators
As I'm reading more of the documentation I'm realizing there are more
than just a couple of parameter types. Here are a few examples:

- Little-endian 8-bit, 16-bit or 32-bit number
- Variable-length null-terminated buffer with hardcoded max size e.g.
  <https://download4.epson.biz/sec_pubs/pos/reference_en/escpos/esc_cd.html>
- Series of key-value pairs with total length in bytes specified in a
  header e.g.
  <https://download4.epson.biz/sec_pubs/pos/reference_en/escpos/fs_lparen_ce_fn64.html>
- Buffer of bytes with total length specified in a header e.g.
  <https://download4.epson.biz/sec_pubs/pos/reference_en/escpos/gs_lparen_cl_fn112.html>

This was previously handled by putting some data in the reflection
metadata and then essentially having a bunch of type-switches in the
CmdBase parent class to handle the different parsing/serialization/
validation modalities. This wasn't going to scale.

In this change I've offloaded the parsing, serialization and validation
logic into parameterized functions stored on the decorators themselves,
to which the CmdBase parent class can delegate the specifics.