Make constructor and validation required in tests
This is to prevent forgetting to call this.validate() at the end of
the subclass constructor.

Also, to reduce duplication, the `class` member of `TestCase` was
removed as it is redundant with the now-required `constructed` (renamed
to `cmd`) member. In lieu of using `class.name` for test names, we are
now using the string description as copied from the documentation
website. To this end, all stubs have had their string descriptions
populated. New stubs going forward should have their string descriptions
populated.