Separate tsconfig just for eslint
tsc should not run on unit test files, lest we distribute our unit
tests to clients, so the main tsconfig should exclude unit test files.
However, this causes eslint to complain when it goes to lint the unit
test files:

    <unit test file> was not found by the project service. Consider
    either including it in the tsconfig.json or including it in
    allowDefaultProject

To fix this, a separate tsconfig just for eslint is added that includes
the unit test files, but isn't used by tsc.