Switch from eslint-specific tsconfig to build-specific tsconfig
I realized there are actually three different clients of tsconfigs:

1. Typescript LSPs
2. Typescript eslint
3. Typescript compiler

Typescript LSPs and eslint want the same thing: they both want to work
on all files. The Typescript compiler wants a slightly different
tsconfig: it wants to exclude unit test files, custom test reporters;
basically anything that's not a source file.

It's easier for the LSP and eslint to rely on the default toplevel
tsconfig.json, so that's what both of those are using now. The build
step pulls in the special tsconfig.build.json, which extends the default
toplevel tsconfig.json, but crucially excludes unit test files and
custom test reporters.